Se adiciona logica para el metodo de eliminar

+++ src/main/java/mx/com/cuh/cuh/service/UsuarioImpl.java
@@ -1,10 +1,12 @@
 package mx.com.cuh.cuh.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import mx.com.cuh.cuh.dto.RespuestaEliminar;
 import mx.com.cuh.cuh.entity.TbPerson;
 import mx.com.cuh.cuh.repository.TbPersonRepository;
 
@@ -20,9 +22,18 @@ public class UsuarioImpl implements Usuario {
 	}
 
 	@Override
-	public void borrarPersona(Long idPerson) {
-		tbPersonRepository.deleteById(idPerson);
+	public RespuestaEliminar borrarPersona(Long idPerson) {
+		Optional<TbPerson> persona =
+				tbPersonRepository.findById(idPerson);
+		RespuestaEliminar response = new RespuestaEliminar();
+		if (persona.isPresent()) {
+			tbPersonRepository.deleteById(idPerson);
+			response.setMensaje("Se eliminó correctamente");
+		}else {
+			response.setMensaje("El usuario " + idPerson + " no existe");
+		}
 		
+		return response;
 	}
